Implement redux on the form to add interactions
