+++ package.json
@@ -84,7 +84,7 @@
     "core-js": "^2.5.7",
     "css-loader": "^1.0.0",
     "csurf": "^1.9.0",
-    "data-hub-components": "^5.3.4",
+    "data-hub-components": "^5.4.0",
     "date-fns": "^1.29.0",
     "del-cli": "^3.0.0",
     "dotenv": "^5.0.0",
