+++ src/client/actions.js
@@ -45,3 +45,17 @@ export const EXPORT_WINS__LOADED = 'EXPORT_WINS__LOADED'
 export const EXPORT_WINS__SELECT_PAGE = 'EXPORT_WINS__SELECT_PAGE'
 
 export const EXPORT_COUNTRIES_EDIT__SAVE = 'EXPORT_COUNTRIES_EDIT__SAVE'
+
+export const FORM__FIELD_SET_VALUE = 'FORM__FIELD_SET_VALUE'
+export const FORM__FIELD_TOUCHED = 'FORM__FIELD_TOUCHED'
+
+export const FORM__FIELD_REGISTER = 'FORM__FIELD_REGISTER'
+export const FORM__FIELD_DEREGISTER = 'FORM__FIELD_DEREGISTER'
+
+export const FORM__STEP_REGISTER = 'FORM__STEP_REGISTER'
+export const FORM__STEP_DEREGISTER = 'FORM__STEP_DEREGISTER'
+
+export const FORM__FORWARD = 'FORM__FORWARD'
+export const FORM__BACK = 'FORM__BACK'
+
+export const FORM__VALIDATE = 'FORM__VALIDATE'
